requirements.txt
Django>=4.2
djangorestframework
django-filter
django-simple-history
django-import-export
psycopg2-binary
gunicorn